+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Expose bias flag for feedforwards, same default as Timm [#220]
 - Update eps value for layernormm, same default as torch [#221]
 - PreNorm bugfix, only one input was normalized [#233]
+- Fix bug where embedding dimensions that did not match model dim would lead to a crash [#244]
 
 ### Added
 - Add DeepNet (DeepNorm) residual path and init [#227]